fix: design fixes
