+++ src/app/[locale]/components/common/metrics-cards/metrics-cards-modal.tsx
@@ -31,7 +31,8 @@ const MetricsCardsModal: FC<OwnProps> = ({ item, title, list, plusButtonSize = '
         opened={isModalOpened}
         hideClose
         onClose={() => setIsModalOpened(false)}
-        className="fixed left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 transform"
+        isRelative
+        className="fixed left-1/2 top-1/2 -translate-x-1/2 -translate-y-4 transform"
       >
         <div className="flex w-fit max-w-[90vw] flex-row flex-wrap">
           <div
